requirements.txt
asyncpg
sqlalchemy[asyncio]
langchain
langchain-groq
pydantic
dotenv
argon2-cffi
python-jose
fastapi
uvicorn
langchain-tavily
langchain-ollama
langchain-pinecone
pinecone
langchain_text_splitters
langchain_community
langchain_classic
python-multipart
pypdf
docx2txt
pillow
pytesseract
pydantic[email]
alembic
psycopg2-binary
apscheduler
google-genai